# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a projectc-ethereum-connector.",
        "title": "Swagger projectc-ethereum-connector",
        "contact": {
            "name": "guyuxiang",
            "url": "https://guyuxiang.github.io",
            "email": "guyuxiang@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/inner/chain-data-subscribe/{networkCode}/address-subscribe": {
            "post": {
                "description": "Register an address scan subscription for callback processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address subscription payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddressSubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data-subscribe/{networkCode}/address-subscribe-cancel": {
            "post": {
                "description": "Update or stop an address scan subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Cancel address subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address subscription cancel payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddressSubscribeCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data-subscribe/{networkCode}/tx-subscribe": {
            "post": {
                "description": "Register a transaction subscription for callback processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction subscription payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TxSubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data-subscribe/{networkCode}/tx-subscribe-cancel": {
            "post": {
                "description": "Cancel a registered transaction subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Cancel transaction subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction subscription cancel payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TxSubscribeCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/address-balance": {
            "post": {
                "description": "Query native token balance of an address on the configured network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Query address balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Address balance query payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddressBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/latest-block": {
            "post": {
                "description": "Query latest block number and timestamp from the configured network",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Query latest block",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/token-add": {
            "post": {
                "description": "Add a token definition into database-backed EVM token registry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Add or update token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Token add request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TokenAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/token-balance": {
            "post": {
                "description": "Query token balance of an address for a token configured in the database-backed EVM token registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Query token balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Token balance query payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TokenBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/token-delete": {
            "post": {
                "description": "Delete a token definition from database-backed EVM token registry by token code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Delete token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Token delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TokenDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/inner/chain-data/{networkCode}/common/token-get": {
            "post": {
                "description": "Get a token definition from database-backed EVM token registry by token code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Configured network code",
                        "name": "networkCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Token get request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TokenGetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddressBalanceRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "models.AddressSubscribeCancelRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "endBlockNumber": {
                    "type": "integer"
                }
            }
        },
...
        "models.AddressSubscribeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "subscribeRange": {
                    "$ref": "#/definitions/models.AddressSubscribeRange"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.TokenAddRequest": {
            "type": "object",
            "properties": {
                "tokenAddress": {
                    "type": "string"
                },
                "tokenCode": {
                    "type": "string"
                }
            }
        },
...
        "models.TxSubscribeCancelRequest": {
            "type": "object",
            "properties": {
                "txCode": {
                    "type": "string"
                }
            }
        },
...
        "models.TxSubscribeRequest": {
            "type": "object",
            "properties": {
                "subscribeRange": {
                    "$ref": "#/definitions/models.TxSubscribeRange"
                },
                "txCode": {
                    "type": "string"
                }
            }
        }
    }
}